+++ dislib/optimization/__init__.py
@@ -1,0 +1,3 @@
+from dislib.optimization.admm.base import ADMM
+
+__all__ = ['ADMM']
