+++ dislib/regression/__init__.py
@@ -1,3 +1,4 @@
 from dislib.regression.linear.base import LinearRegression
+from dislib.regression.lasso.base import Lasso
 
-__all__ = ['LinearRegression']
+__all__ = ['LinearRegression', 'Lasso']
