+++ examples/admm.py
@@ -1,0 +1,25 @@
+import sys
+import time
+
+from dislib.optimization import ADMM
+from dislib.regression import Lasso
+
+
+def main():
+    start = time.time()
+    n = int(sys.argv[1])
+
+    optimizer = ADMM(rho=1, abstol=1e-4, reltol=1e-2)
+    lasso = Lasso(n=n, max_iter=500, lmbd=1e-3, optimizer=optimizer)
+    optimizer.objective_fn = lasso.objective_x
+
+    z = lasso.fit()
+
+    print("\nTotal elapsed time: %s" % str((time.time() - start) / 100))
+    np.savetxt("Solution.COMPSs.txt", z)
+
+    print(lasso.predict(np.random.rand(50, 50)))
+
+
+if __name__ == '__main__':
+    main()
