+++ requirements.txt
@@ -2,3 +2,4 @@ scikit-learn>=0.19.1
 scipy>=1.3.0
 numpy>=1.15.4
 numpydoc>=0.8.0
+cvxpy
